+++ src/main/java/com/mtf/admin/entity/Agency.java
@@ -24,7 +24,7 @@ public class Agency implements Serializable {
     private String phone;           //电话
     private Integer roomCard;       //房卡
     private Integer coin;           //金币
-    private Integer agencyBalance;  //代理余额
+    private Double agencyBalance;  //代理余额
     private Integer disable;        //禁用
     private Integer deleted;        //删除
     private Date createDate;        //创建时间
